feat: [arieftb] Add New Note Page
move the section of add new note from home to a new specific page
