+++ src/presentation/templates/NotesTemplate.jsx
@@ -1,25 +1,19 @@
 import { NoteSearch } from '../molecules/NoteSearch';
-import { NoteForm } from '../molecules/NoteForm';
 import { NoteList } from '../organisms/NoteList';
 import PropTypes from 'prop-types';
+import { Link } from 'react-router-dom';
 
 export const NotesLayout = ({
   activeNotes,
   archivedNotes,
   searchQuery,
   onSearchChange,
-  onAddNote,
   onDeleteNote,
   onToggleArchive,
   onToggleUnArchive,
 }) => {
   return (
     <div className="container">
-      <section className="mb-5">
-        <h2 className="mb-3">Add New Note</h2>
-        <NoteForm onSubmit={onAddNote}/>
-      </section>
-
       <section className="mb-4">
         <NoteSearch value={searchQuery} onChange={onSearchChange}/>
       </section>
@@ -47,6 +41,10 @@ export const NotesLayout = ({
           />
         </div>
       </section>
+
+      <Link to="/notes/new" className="floating-button">
+        <span className="plus-icon">+</span>
+      </Link>
     </div>
   );
 };
@@ -56,7 +54,6 @@ NotesLayout.propTypes = {
   archivedNotes: PropTypes.array.isRequired,
   searchQuery: PropTypes.string.isRequired,
   onSearchChange: PropTypes.func.isRequired,
-  onAddNote: PropTypes.func.isRequired,
   onDeleteNote: PropTypes.func.isRequired,
   onToggleArchive: PropTypes.func.isRequired,
   onToggleUnArchive: PropTypes.func.isRequired,
